+++ backend/docs/test.yaml
@@ -1,0 +1,8 @@
+definitions:
+  LoginCredentials:
+    type: "object"
+    properties:
+      email:
+        type: "string"
+      password:
+        type: "string"
